refactor project to use iuvo 0.0.3
