+++ src/main/java/dev/sirtimme/scriletio/commands/interaction/component/button/AcceptDeletionButton.java
@@ -1,12 +1,11 @@
 package dev.sirtimme.scriletio.commands.interaction.component.button;
 
-import dev.sirtimme.iuvo.commands.interaction.IInteractionCommand;
-import dev.sirtimme.iuvo.precondition.IPrecondition;
-import dev.sirtimme.iuvo.repository.QueryableRepository;
-import dev.sirtimme.iuvo.repository.Repository;
+import dev.sirtimme.iuvo.api.commands.interaction.IInteractionCommand;
+import dev.sirtimme.iuvo.api.precondition.IPrecondition;
+import dev.sirtimme.iuvo.api.repository.QueryableRepository;
+import dev.sirtimme.iuvo.api.repository.Repository;
 import dev.sirtimme.scriletio.entities.DeleteConfig;
 import dev.sirtimme.scriletio.entities.User;
-import dev.sirtimme.scriletio.precondition.component.IsComponentAuthor;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 
 import java.util.Collections;
@@ -36,7 +35,7 @@ public class AcceptDeletionButton implements IInteractionCommand<ButtonInteracti
     @Override
     public List<IPrecondition<? super ButtonInteractionEvent>> getPreconditions() {
         return List.of(
-            new IsComponentAuthor()
+            IPrecondition.isComponentAuthor()
         );
     }
 }
