+++ src/main/java/dev/sirtimme/scriletio/commands/interaction/component/button/AcceptRegistrationButton.java
@@ -1,10 +1,9 @@
 package dev.sirtimme.scriletio.commands.interaction.component.button;
 
-import dev.sirtimme.iuvo.commands.interaction.IInteractionCommand;
-import dev.sirtimme.iuvo.precondition.IPrecondition;
-import dev.sirtimme.iuvo.repository.Repository;
+import dev.sirtimme.iuvo.api.commands.interaction.IInteractionCommand;
+import dev.sirtimme.iuvo.api.precondition.IPrecondition;
+import dev.sirtimme.iuvo.api.repository.Repository;
 import dev.sirtimme.scriletio.entities.User;
-import dev.sirtimme.scriletio.precondition.component.IsComponentAuthor;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 
 import java.util.Collections;
@@ -28,7 +27,7 @@ public class AcceptRegistrationButton implements IInteractionCommand<ButtonInter
     @Override
     public List<IPrecondition<? super ButtonInteractionEvent>> getPreconditions() {
         return List.of(
-            new IsComponentAuthor()
+            IPrecondition.isComponentAuthor()
         );
     }
 }
