+++ src/main/java/dev/sirtimme/scriletio/commands/interaction/component/button/CancelDeletionButton.java
@@ -1,8 +1,7 @@
 package dev.sirtimme.scriletio.commands.interaction.component.button;
 
-import dev.sirtimme.iuvo.commands.interaction.IInteractionCommand;
-import dev.sirtimme.iuvo.precondition.IPrecondition;
-import dev.sirtimme.scriletio.precondition.component.IsComponentAuthor;
+import dev.sirtimme.iuvo.api.commands.interaction.IInteractionCommand;
+import dev.sirtimme.iuvo.api.precondition.IPrecondition;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 
 import java.util.Collections;
@@ -18,7 +17,7 @@ public class CancelDeletionButton implements IInteractionCommand<ButtonInteracti
     @Override
     public List<IPrecondition<? super ButtonInteractionEvent>> getPreconditions() {
         return List.of(
-            new IsComponentAuthor()
+            IPrecondition.isComponentAuthor()
         );
     }
 }
