+++ src/main/java/dev/sirtimme/scriletio/commands/interaction/sub/DeleteConfigCommand.java
@@ -1,13 +1,12 @@
 package dev.sirtimme.scriletio.commands.interaction.sub;
 
-import dev.sirtimme.iuvo.commands.interaction.ISubCommand;
-import dev.sirtimme.iuvo.precondition.IPrecondition;
-import dev.sirtimme.iuvo.repository.QueryableRepository;
-import dev.sirtimme.iuvo.repository.Repository;
+import dev.sirtimme.iuvo.api.commands.interaction.ISubCommand;
+import dev.sirtimme.iuvo.api.precondition.IPrecondition;
+import dev.sirtimme.iuvo.api.repository.QueryableRepository;
+import dev.sirtimme.iuvo.api.repository.Repository;
 import dev.sirtimme.scriletio.entities.DeleteConfig;
 import dev.sirtimme.scriletio.entities.User;
-import dev.sirtimme.scriletio.precondition.slash.HasSavedConfigs;
-import dev.sirtimme.scriletio.precondition.slash.IsRegistered;
+import dev.sirtimme.scriletio.precondition.HasSavedConfigs;
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import net.dv8tion.jda.api.entities.emoji.Emoji;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
@@ -59,7 +58,7 @@ public class DeleteConfigCommand implements ISubCommand {
     public List<IPrecondition<? super SlashCommandInteractionEvent>> getPreconditions() {
         return List.of(
             new HasSavedConfigs(configRepository),
-            new IsRegistered(userRepository)
+            IPrecondition.isRegistered(userRepository)
         );
     }
