+++ src/main/java/dev/sirtimme/scriletio/factory/interaction/SlashEventCommandFactory.java
@@ -1,8 +1,8 @@
 package dev.sirtimme.scriletio.factory.interaction;
 
-import dev.sirtimme.iuvo.commands.interaction.IInteractionCommand;
-import dev.sirtimme.iuvo.commands.interaction.ISlashCommand;
-import dev.sirtimme.iuvo.factory.interaction.IInteractionCommandFactory;
+import dev.sirtimme.iuvo.api.commands.interaction.IInteractionCommand;
+import dev.sirtimme.iuvo.api.commands.interaction.ISlashCommand;
+import dev.sirtimme.iuvo.api.factory.interaction.IInteractionCommandFactory;
 import dev.sirtimme.scriletio.commands.interaction.slash.*;
 import dev.sirtimme.scriletio.repository.UserRepository;
 import dev.sirtimme.scriletio.repository.DeleteConfigRepository;
