+++ src/main/java/dev/sirtimme/scriletio/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package dev.sirtimme.scriletio.repository;
 
-import dev.sirtimme.iuvo.repository.Repository;
+import dev.sirtimme.iuvo.api.repository.Repository;
 import dev.sirtimme.scriletio.entities.User;
 import jakarta.persistence.EntityManager;
